change to double

+++ src/main/java/com/example/springdocker/service/MyMathCalc.java
@@ -15,7 +15,7 @@ public class MyMathCalc {
     public int multiply(int a, int b){
         return a*b;
     }
-    public int divide(int a, int b){
+    public double divide(int a, int b){
         return a/b;
     }
 }
